incrementGeneratedGroup: support progressive loading of groups
This could likely be improved so that only non-dynamic groups get progressively loaded -- as it is, if we have 7 groups, and two are dynamic AJAX calls, then we would expect the layout to be drawn 5 times immediately, then a sixth time when the first AJAX call resolves, and then a seventh time when the second AJAX call resolves. Ideally, we would only do it three times.